# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Xeno Sir AI Tutor
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login as admin or student
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    username: zod.string(),
    role: zod.enum(["admin", "student"]),
    displayName: zod.string(),
  }),
  token: zod.string(),
});

/**
 * @summary Get current user info
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  role: zod.enum(["admin", "student"]),
  displayName: zod.string(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary List all lectures
 */
export const ListLecturesResponseItem = zod.object({
  id: zod.number(),
  lectureNumber: zod.number(),
  title: zod.string(),
  transcript: zod.string(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListLecturesResponse = zod.array(ListLecturesResponseItem);

/**
 * @summary Create/add a new lecture transcript (admin only)
 */
export const CreateLectureBody = zod.object({
  lectureNumber: zod.number(),
  title: zod.string(),
  transcript: zod.string(),
});

/**
 * @summary Get a single lecture
 */
export const GetLectureParams = zod.object({
  id: zod.coerce.number(),
});

export const GetLectureResponse = zod.object({
  id: zod.number(),
  lectureNumber: zod.number(),
  title: zod.string(),
  transcript: zod.string(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Update a lecture (admin only)
 */
export const UpdateLectureParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateLectureBody = zod.object({
  lectureNumber: zod.number().optional(),
  title: zod.string().optional(),
  transcript: zod.string().optional(),
});

export const UpdateLectureResponse = zod.object({
  id: zod.number(),
  lectureNumber: zod.number(),
  title: zod.string(),
  transcript: zod.string(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete a lecture (admin only)
 */
export const DeleteLectureParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all students (admin only)
 */
export const ListStudentsResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  displayName: zod.string(),
  createdAt: zod.date(),
});
export const ListStudentsResponse = zod.array(ListStudentsResponseItem);

/**
 * @summary Create a new student account (admin only)
 */
export const CreateStudentBody = zod.object({
  username: zod.string(),
  password: zod.string(),
  displayName: zod.string(),
});

/**
 * @summary Get student details with chat history (admin only)
 */
export const GetStudentParams = zod.object({
  id: zod.coerce.number(),
});

export const GetStudentResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  displayName: zod.string(),
  createdAt: zod.date(),
  conversations: zod.array(
    zod.object({
      id: zod.number(),
      title: zod.string(),
      createdAt: zod.date(),
    }),
  ),
});

/**
 * @summary Delete a student (admin only)
 */
export const DeleteStudentParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary Ask Xeno Sir a question with lecture context
 */
export const AskXenoSirBody = zod.object({
  question: zod.string(),
  lectureIds: zod
    .array(zod.number())
    .optional()
    .describe(
      "Optional - specific lecture IDs to focus on. If empty, uses all lectures.",
    ),
  conversationId: zod
    .number()
    .optional()
    .describe("Optional - continue an existing conversation"),
  language: zod
    .enum(["hindi", "english", "auto"])
    .optional()
    .describe("Response language preference"),
});

/**
 * @summary Generate an exam/quiz from lecture transcripts
 */
export const generateExamBodyQuestionCountMin = 5;
export const generateExamBodyQuestionCountMax = 50;

export const GenerateExamBody = zod.object({
  lectureIds: zod
    .array(zod.number())
    .optional()
    .describe(
      "Lecture IDs to generate exam from. If empty, uses all lectures.",
    ),
  difficulty: zod.enum(["basic", "medium", "advanced", "extreme"]),
  examType: zod.enum(["mcq", "subjective", "mixed"]),
  questionCount: zod
    .number()
    .min(generateExamBodyQuestionCountMin)
    .max(generateExamBodyQuestionCountMax),
});

export const GenerateExamResponse = zod.object({
  lecturesCovered: zod.array(zod.string()),
  difficulty: zod.string(),
  questions: zod.array(
    zod.object({
      questionNumber: zod.number(),
      question: zod.string(),
      type: zod.enum(["mcq", "subjective"]),
      options: zod.array(zod.string()).optional().describe("For MCQ only"),
      correctAnswer: zod.string(),
      explanation: zod.string(),
    }),
  ),
  totalQuestions: zod.number(),
});

/**
 * @summary Get chat history for a student (admin only)
 */
export const GetChatHistoryParams = zod.object({
  studentId: zod.coerce.number(),
});

export const GetChatHistoryResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
});
export const GetChatHistoryResponse = zod.array(GetChatHistoryResponseItem);