# lib/api-zod/src/generated/types/createStudentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Xeno Sir AI Tutor
 * OpenAPI spec version: 0.1.0
 */

export interface CreateStudentBody {
  username: string;
  password: string;
  displayName: string;
}