# lib/api-zod/src/generated/types/studentWithConversations.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Xeno Sir AI Tutor
 * OpenAPI spec version: 0.1.0
 */
import type { OpenaiConversation } from "./openaiConversation";

export interface StudentWithConversations {
  id: number;
  username: string;
  displayName: string;
  createdAt: Date;
  conversations: OpenaiConversation[];
}